Revert "Bring the DemoActivity to the front on deep link click, even when the app process is killed"
This reverts the previous commit.

The previous commit didn't work because:
As soon as `BrowserSwitchActivity` finishes itself, the OS kills the process. This means that the lifecycle callbacks listener is never notified, so DemoActivity is never brought to the front. The only way to go back to DemoActivity is to close custom tabs manually.