Replace the activity lifecycles callbacks with a static field in BrowserSwitchActivity
We need a solution for `DemoActivity` to be launched when `BrowserSwitchActivity` finishes.

Registering an `ActivityLifecycleCallback` looks cleaner, as it's an api provided explicitly to know when different activities are destroyed. However, when you set the developer options to 1) don't keep activities and 2) have no background processes, it's not possible to have this `ActivityLifecycleCallback` registered: When you tap on a link in custom tabs, `BrowserSwitchActivity` is created and finishes itself, and no lifecycle callback is there to listen to this. Even if you register a lifecycle callback from the application, this doesn't work: As soon as `BrowserSwitchActivity` finishes itself, Android kills the process, and the lifecycle callback isn't notified.

Telling `BrowserActivity` which intent it needs to launch when it finishes itself solves this problem. We can do this from the `Application.onCreate()` (this is done in a following commit)